requirements.txt
fastapi
uvicorn[standard]
dvc
gitpython
hydra-core
pydantic
python-multipart
